ex17_js_components are done.
