+++ src/kanban/index.css
@@ -1,0 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
+@import '/src/kanban/src/style/header.css';
+@import '/src/kanban/src/style/main.css';
+@import '/src/kanban/src/style/content.css';
+@import '/src/kanban/src/style/footer.css';
