+++ src/kanban/index.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="task-03.css">
+  <link rel="stylesheet" href="index.css">
   <title>Task-03</title>
 </head>
 
@@ -152,8 +152,6 @@
       </div>
     </footer>
   </main>
-  <script src="events.js"></script>
-  <script src="dropdown.js"></script>
+  <script type="module" src="index.js"></script>
 </body>
-
 </html>
