+++ src/kanban/index.js
@@ -1,0 +1,5 @@
+import {dropdpwn} from './src/script/dropdown.js';
+dropdpwn();
+
+import {renderTasks} from './src/script/events.js';
+renderTasks();
