+++ src/kanban/src/script/eventsMock.js
@@ -1,0 +1,79 @@
+export const LS_NAME_TASKS = 'tasks';
+export const initialData = [
+  {
+    title: 'Backlog',
+    tasks: [
+      {
+        id: 'task0',
+        name: 'Login page - performance issues'
+      },
+      {
+        id: 'task1',
+        name: 'Sprint bugfix'
+      }
+    ],
+  },
+  {
+    title: 'Ready',
+    tasks: [
+      {
+        id: 'task2',
+        name: 'Shop page - performance issues'
+      },
+      {
+        id: 'task3',
+        name: 'Checkout bugfix'
+      },
+      {
+        id: 'task4',
+        name: 'Shop bug 1'
+      },
+      {
+        id: 'task5',
+        name: 'Shop bug 2'
+      },
+      {
+        id: 'task6',
+        name: 'Shop bug 3'
+      },
+      {
+        id: 'task7',
+        name: 'Shop bug 4'
+      },
+      {
+        id: 'task8',
+        name: 'Shop bug 5'
+      },
+      {
+        id: 'task9',
+        name: 'Shop bug 6'
+      }
+    ],
+  },
+  {
+    title: 'in progress',
+    tasks: [
+      {
+        id: 'task10',
+        name: 'User page - performance issues'
+      },
+      {
+        id: 'task11',
+        name: 'Auth bugfix'
+      }
+    ],
+  },
+  {
+    title: 'Finished',
+    tasks: [
+      {
+        id: 'task12',
+        name: 'Main page - performance issues'
+      },
+      {
+        id: 'task13',
+        name: 'Main page bugfix'
+      }
+    ],
+  },
+]
