+++ src/kanban/src/script/eventsVariable.js
@@ -1,0 +1,29 @@
+export const blocks = {
+  backlog: document.querySelector('.backlog-tasks'),
+  ready: document.querySelector('.ready-tasks'),
+  inProgress: document.querySelector('.inProgress-tasks'),
+  finished: document.querySelector('.finished-tasks'),
+};
+
+
+ 
+export const addButtons = {
+  backlog: document.querySelector('.backlog-add'),
+  ready: document.querySelector('.ready-add'),
+  inProgress: document.querySelector('.inProgress-add'),
+  finished: document.querySelector('.finished-add'),
+};
+
+
+export const selectButtons = {
+  ready: document.querySelector('.ready-select'),
+  inProgress: document.querySelector('.inProgress-select'),
+  finished: document.querySelector('.finished-select'),
+};
+
+
+export const listButton = {
+  ready: document.querySelector('.ready-list'),
+  inProgress: document.querySelector('.inProgress-list'),
+  finished: document.querySelector('.finished-list'),
+};
