+++ src/kanban/src/style/content-box-content.css
@@ -1,0 +1,10 @@
+.box-content{
+  display: flex;
+  position: relative;
+  background-color:#EBECF0;
+  width: 100%;
+  max-height: 70vh;
+  align-self: center;
+  flex-direction: column;
+  overflow: auto;
+}
