+++ src/kanban/src/style/content-box-footer.css
@@ -1,0 +1,20 @@
+.box-footer{
+  display: flex;
+  position: relative;
+  height: 35px;
+  width: 100%;
+  background-color:  #EBECF0;
+  border-radius: 15px;
+  cursor: default;
+}
+
+.box-footer h4{
+  position: absolute;
+  font-family: "Roboto";
+  font-style: normal;
+  font-weight: normal;
+  font-size: 18px;
+  line-height: 21px;
+  margin-left: 5px;
+  color: #5E6C84;
+}
