+++ src/kanban/src/style/content-box.css
@@ -1,0 +1,13 @@
+.box{
+  display: flex;
+  position: relative;
+  width: 282px;
+  background: #EBECF0;
+  margin: 20px;
+  border: 1px solid #EBECF0;
+  border-radius: 10px;
+  flex-direction: column;
+  justify-self: start;
+  align-self: flex-start;
+  max-height: 100vh;
+}
