+++ src/kanban/src/style/content.css
@@ -1,0 +1,18 @@
+@import '/src/kanban/src/style/drop-down.css';
+@import '/src/kanban/src/style/content-box.css';
+@import '/src/kanban/src/style/content-box-header.css';
+@import '/src/kanban/src/style/content-box-content.css';
+@import '/src/kanban/src/style/content-box-footer.css';
+@import '/src/kanban/src/style/content-box-events.css';
+
+.content{
+  display: flex;
+  width: 100%;
+  height: 100%;
+  background-color: #0079BF;
+  position: relative;
+  align-items: flex-end;
+  justify-content: center;
+  max-height: 100vh;
+  overflow: hidden;
+}
