+++ src/kanban/src/style/create-new-list.css
@@ -1,0 +1,29 @@
+.create-button{
+  display: flex;
+  align-items: center;
+}
+
+.create{
+  display: flex;
+  position: absolute;
+  width: 181px;
+  height: 33px;
+  background: #FFFFFF;
+  border-radius: 5px;
+  align-items: center;
+  justify-content: center;
+  font-family: "Roboto";
+  font-style: normal;
+  font-weight: normal;
+  font-size: 18px;
+  line-height: 21px;
+  border: none;
+  outline: none;
+  cursor: pointer;
+}
+
+.black-plus{
+  position: relative;
+  align-self: center;
+  right: 8px;
+}
