+++ src/kanban/src/style/drop-down.css
@@ -1,0 +1,40 @@
+.drop-div{
+  content: "";
+  position: absolute;
+  display: flex;
+  height: 130px;
+  width: 140px;
+  background-color: #ffffff;
+  top: 55px;
+  margin-left: 1595px;
+  border: 1px solid #ffffff;
+  border-radius: 15px;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.line{
+  position: relative;
+  display: flex;
+  height: 35px;
+  width: 140px;
+  align-self: center;
+  background-color: rgb(255, 255, 255);
+  flex-direction: column;
+  justify-self: space-around;
+  font-size: 20px;
+  align-items: flex-start;
+  justify-content: center;
+  font-family: "Roboto";
+  font-style: normal;
+  font-weight: normal;
+  font-size: 18px;
+  line-height: 21px;
+  cursor: pointer;
+  border: 0;
+  outline: none;
+}
+
+.line:hover{
+  background-color: #0067a327;
+}
