+++ src/kanban/src/style/footer-boxes.css
@@ -1,0 +1,14 @@
+.firsit-box{
+  position: relative;
+  margin-left: 70px;
+}
+
+.second-box{
+  position: relative;
+  right: 120px;
+}
+
+.third-box{
+  position: relative;
+  left: 680px;
+}
