+++ src/kanban/src/style/footer.css
@@ -1,0 +1,31 @@
+@import '/src/kanban/src/style/footer-boxes.css';
+
+.footer{
+  display: flex;
+  position: relative;
+  width: 100%;
+  height: 55px;
+  background: #0067A3;
+  flex-direction: row;
+  align-items: center;
+  font-family: "Roboto";
+  font-style: normal;
+  font-weight: normal;
+  font-size: 18px;
+  line-height: 21px;
+  color: #ffffff;
+  flex: 0 0 auto;
+  align-self: flex-end;
+}
+
+input{
+  outline: none;
+  border: none;
+  background: transparent;
+  color: #ffffff;
+  font-family: "Roboto";
+  font-style: normal;
+  font-weight: normal;
+  font-size: 18px;
+  line-height: 21px;
+}
