--- src/kanban/task-03.css
@@ -1,433 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
-
-*{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-.main{
-  display: flex;
-  width: 100%;
-  height: 100vh;
-  position: relative;
-  flex-direction: column;
-  
-  
-}
-
-.header{
-  display: flex;
-  position: relative;
-  flex-direction: row;
-  width: 100%;
-  height: 55px;
-  background: #0067A3;
-  flex: 0 0 auto;
-  align-self: flex-start;
-}
-
-.logo{
-  margin-left: 25px;
-  align-self: center;
-  margin-top: 5px;
-}
-
-.logo img{
-  height: 45px;
-  width: 45px;
-}
-
-.logo h1{
-  width: 331px;
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: normal;
-  font-size: 28px;
-  color: #FFFFFF;
-}
-
-.create-button{
-  display: flex;
-  align-items: center;
-}
-
-.create{
-  display: flex;
-  position: absolute;
-  width: 181px;
-  height: 33px;
-  background: #FFFFFF;
-  border-radius: 5px;
-  align-items: center;
-  justify-content: center;
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  border: none;
-  outline: none;
-  cursor: pointer;
-}
-
-.right-menu{
-  display: flex;
-  flex-direction: row;
-  position: relative;
-  margin-left: 1020px;
-  align-items: center;
-  align-self: center;
-}
-
-.white-img{
-  display: flex;
-  align-self: center;
-  position: relative;
-  margin-top: 5px;
-  margin-left: 230px;
-}
-
-.black-img{
-  position: absolute;
-  align-self: center;
-}
-
-.arrow{
-  position: relative;
-  margin-left: 5px;
-  display: flex;
-}
-
-.black-plus{
-  position: relative;
-  align-self: center;
-  right: 8px;
-}
-
-.footer{
-  display: flex;
-  position: relative;
-  width: 100%;
-  height: 55px;
-  background: #0067A3;
-  flex-direction: row;
-  align-items: center;
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  color: #ffffff;
-  flex: 0 0 auto;
-  align-self: flex-end;
-}
-
-.firsit-box{
-  position: relative;
-  margin-left: 70px;
-}
-
-.second-box{
-  position: relative;
-  right: 120px;
-}
-
-.third-box{
-  position: relative;
-  left: 680px;
-}
-
-input{
-  outline: none;
-  border: none;
-  background: transparent;
-  color: #ffffff;
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-}
-
-
-.drop-div{
-  content: "";
-  position: absolute;
-  display: flex;
-  height: 130px;
-  width: 140px;
-  background-color: #ffffff;
-  top: 55px;
-  margin-left: 1595px;
-  border: 1px solid #ffffff;
-  border-radius: 15px;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.line{
-  position: relative;
-  display: flex;
-  height: 35px;
-  width: 140px;
-  align-self: center;
-  background-color: rgb(255, 255, 255);
-  flex-direction: column;
-  justify-self: space-around;
-  font-size: 20px;
-  align-items: flex-start;
-  justify-content: center;
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  cursor: pointer;
-  border: 0;
-  outline: none;
-}
-
-.line:hover{
-  background-color: #0067a327;
-}
-
-.content{
-  display: flex;
-  width: 100%;
-  height: 100%;
-  background-color: #0079BF;
-  position: relative;
-  align-items: flex-end;
-  justify-content: center;
-  max-height: 100vh;
-  overflow: hidden;
-}
-
-.box{
-  display: flex;
-  position: relative;
-  width: 282px;
-  background: #EBECF0;
-  margin: 20px;
-  border: 1px solid #EBECF0;
-  border-radius: 10px;
-  flex-direction: column;
-  justify-self: start;
-  align-self: flex-start;
-  max-height: 100vh;
-}
-
-.box-header{
-  display: flex;
-  position: relative;
-  height: 38px;
-  width: 100%;
-  background-color:  #EBECF0;
-  border-radius: 15px;
-  justify-content: space-between;
-  cursor: default;
-}
-.heder-txt{
-  display: flex;
-  align-items: center;
-  margin-left: 5px;
-}
-
-.dots{
-  display: flex;
-  align-items: center;
-  height: 18px;
-  width: 20px;
-  align-self: center;
-  margin-right: 5px;
-  font-size: 20px;
-  border: none;
-  outline: none;
-  background-color: #EBECF0;
-}
-
-.heder-txt h1{
-  position: relative;
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: normal;
-  font-size: 20px;
-  line-height: 21px;
-}
-
-.box-content{
-  display: flex;
-  position: relative;
-  background-color:#EBECF0;
-  width: 100%;
-  max-height: 70vh;
-  align-self: center;
-  flex-direction: column;
-  overflow: auto;
-}
-
-.box-footer{
-  display: flex;
-  position: relative;
-  height: 35px;
-  width: 100%;
-  background-color:  #EBECF0;
-  border-radius: 15px;
-  cursor: default;
-}
-
-.box-footer h4{
-  position: absolute;
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  margin-left: 5px;
-  color: #5E6C84;
-}
-
-.add{
-  display: flex;
-  align-items: center;
-  height: 20px;
-  align-self: center;
-  border: none;
-  outline: none;
-  cursor: pointer;
-}
-
-.add-txt:hover{
-  color: #0079BF;
-}
-
-.item:hover {
-  background-color: #0067a327;
-}
-
-.box .task-list li {
-  background: #ffffff;
-  margin: 0px 5px 10px 5px;
-  padding: 6px 8px;
-  overflow: auto;
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  color: #000000;
-  cursor: default;
-}
-
-.task-input, .select {
-  width: 100%;
-  outline: none;
-  border: none;
-  margin-bottom: 12px;
-  padding: 6px 8px;
-  height: 35px;
-  border-radius: 5px;
-  display: none;
-  background-color: #ffffff;
-  position: relative;
-}
-
-.task-input{
-  height: 10px;
-  color: #000000;
-  max-width: 95%;
-  overflow: auto;
-  align-self: center;
-  padding: 10px;
-  top: 10px;
-}
-
-.select .select-list {
-  text-align: right;
-}
-
-.select-list {
-  background-color: #ffffff;
-  position: relative;
-  list-style: none;
-  left: 12px;
-  margin: 0;
-  padding: 0;
-  padding-top: 5px;
-  bottom: 5px;
-  left: 0px;
-  width: 100%;
-}
-
-.select-list .item {
-  padding: 5px;
-  margin: 0px;
-  font-size: 16px;
-  text-align: left;
-}
-
-.select{
-  border-radius: 15px;
-  border: 1px solid #EBECF0;
-  outline: none;
-  width: 100%;
-  margin: 0px 5px 0px 5px;
-  cursor: default;
-}
-
-.select h1{
-  display: flex;
-  margin-left: 75px;
-  font-size: 20px;
-  color: #5E6C84;
-  cursor: pointer;
-}
-
-.select h1:hover{
-  color:  #0079BF;
-}
-
-.visible {
-  display: block;
-}
-
-.invisible {
-  display: none;
-}
-
-*::-webkit-scrollbar{
-  background-color:#ffffff;
-  width: 12px;
-  height: 12px;
-}
-
-*::-webkit-scrollbar-thumb{
-  background-color:#CBCBCB;
-  border-radius: 5px;
-  width: 12px;
-  height: 52px;
-}
-
-*::-webkit-scrollbar-track{
-  background-color: #ffffff;
-  width: 12px;
-  height: 52px;
-}
-  
-  
-*::-webkit-scrollbar-button{
-  display: none;
-}
-
-.select-list{
-  overflow: auto;
-  max-height: 280px;
-}
-
-.open{
-  background-color: #ffffff;
-  height: 30px;
-  overflow: hidden;
-  border-radius: 15px;
-}
